Query3

+++ data/Queries/src/main/java/Utils/SmartPlugParser.java
@@ -16,7 +16,7 @@ public class SmartPlugParser {
         String[] split = line.split(",");
 
 //        smartPlug = new SmartPlug(Integer.parseInt(split[0]),Integer.parseInt(split[1]),Float.valueOf(split[2]),Integer.parseInt(split[3]),Integer.parseInt(split[4]),Integer.parseInt(split[5]),Integer.parseInt(split[6]));
-        smartPlug = new SmartPlug(Long.parseLong(split[0]),Long.parseLong(split[1]),Float.valueOf(split[2]),Long.parseLong(split[3]),Long.parseLong(split[4]),Long.parseLong(split[5]),Long.parseLong(split[6]));
+        smartPlug = new SmartPlug(Long.parseLong(split[0]),Long.parseLong(split[1]),Double.valueOf(split[2]),Long.parseLong(split[3]),Long.parseLong(split[4]),Long.parseLong(split[5]),Long.parseLong(split[6]));
 
 
         return smartPlug;
@@ -30,7 +30,7 @@ public class SmartPlugParser {
         String[] split = line.split(",");
 
 //        smartPlug = new SmartPlug(Integer.parseInt(split[0]),Integer.parseInt(split[1]),Float.valueOf(split[2]),Integer.parseInt(split[3]),Integer.parseInt(split[4]),Integer.parseInt(split[5]),Integer.parseInt(split[6]));
-        smartPlug = new SmartPlug(Long.parseLong(split[0]),Long.parseLong(split[1]),Float.valueOf(split[2]),Long.parseLong(split[3]),Long.parseLong(split[4]),Long.parseLong(split[5]),Long.parseLong(split[6]));
+        smartPlug = new SmartPlug(Long.parseLong(split[0]),Long.parseLong(split[1]),Double.valueOf(split[2]),Long.parseLong(split[3]),Long.parseLong(split[4]),Long.parseLong(split[5]),Long.parseLong(split[6]));
 
 
         return smartPlug;
